hullModLogic. Implemented BuildInFilter, ShieldTypeFilter, and array split between green and red array
