+++ css/HullMods/hullmods.css
@@ -15,22 +15,25 @@
 	.installed-hullmods {
 		display: flex;
 		flex-direction: column;
-		gap: 2rem;
+		gap: 1.6rem;
 	}
 	.hullmods__container {
 		display: flex;
 		flex-direction: column;
-		gap: 2rem;
+		gap: 1.6rem;
 	}
 	.build-in-hullmods {
 		display: flex;
 		flex-direction: column;
 		margin-left: auto;
 
+		background-color: rgba(0, 0, 0, 0.25);
+		border: 1px solid rgba(0, 0, 0, 0.5);
+
+		padding: 1.6rem;
+
 		.hullmod {
-			background-color: rgba(0, 0, 0, 0.25);
-			border: 1px solid rgba(0, 0, 0, 0.5);
-			width: max-content;
+			/* width: max-content; */
 
 			padding-left: 2.4rem;
 			/* padding: 1.4rem 1.4rem 1.4rem 2.4rem; */
