--- src/starsector_data/HullModData.js
@@ -1,88 +1,0 @@
-// The Data was manually collected by Me.
-// First array is regular data, second array is sMods
-const HULLMODS_DATA = {
-	"Accelerated Shields": [["100%", "100%"], ["100%"]],
-	"Adaptive Phase Coils": [["50%", "50%", "75%"], []],
-	"Automated Repair Unit": [["50%"], ["25%", "33%"]],
-	"Armored Weapon Mounts": [["100%", "25%", "25%", "10%"], ["10%"]],
-	"Advanced Turret Gyros": [["75%"], ["25%", "5%"]],
-	"Additional Berthing": [[30, 60, 100, 200, "30%", "50%"], []],
-	"Auxiliary Fuel Tanks": [[30, 60, 100, 200, "30%", "50%"], []],
-	"Augmented Drive Field": [[2], [1]],
-	"Extended Shields": [[60], [60]],
-	"Hardened Shields": [["20%"], []],
-	"Stabilized Shields": [["50%"], ["10%"]],
-	"Shield Conversion - Front": [["100%"], ["5%"]],
-	"Shield Conversion - Omni": [["30%"], []],
-	"Shield Shunt": [["15%"], ["15%"]],
-	"Makeshift Shield Generator": [[90, "20%"], []],
-	"Flux Shunt": [["50%"], []],
-	"Blast Doors": [["20%", "60%"], ["85%"]],
-	"Heavy Armor": [[150, 300, 400, 500], ["25%"]],
-	"Reinforced Bulkheads": [["40%"], []],
-	"Resistant Flux Conduits": [["50%", "25%"], ["10%"]],
-	"Integrated Point Defense AI": [["50%"], []],
-	"Solar Shielding": [["75%", "10%"], ["100%"]],
-	"Advanced Optics": [[200, "30%"], []],
-	"Ballistic Rangefinder": [[200, 100, 900, 100, 800, 100], []],
-	"Dedicated Targeting Core": [
-		["35%", "50%"],
-		["40%", "60%"],
-	],
-	"ECCM Package": [["50%", "25%", "50%", "25%"], []],
-	"Escort Package": [[1000, "25%", "10%", "20%", "doubled"], ["10%"]],
-	"Expanded Magazines": [["50%"], ["50%"]],
-	"Expanded Missile Racks": [["100%"], ["20%"]],
-	"Missile Autoloader": [[], []],
-	"High Scatter Amplifier": [["10%", 200, "50%"], ["5%"]],
-	"Integrated Targeting Unit": [["10%", "20%", "40%", "60%"], []],
-	"ECM Package": [["1%", "2%", "3%", "4%"], []],
-	"Missile Autoforge": [[], []],
-	"Energy Bolt Coherer": [[100, "50%"], []],
-	"Auxiliary Thrusters": [["50%"], [10]],
-	"Unstable Injector": [[25, 20, 15, 15, "15%", "25%"], []],
-	"Safety Overrides": [[50, 30, 20, 2, 3, 450], []],
-	"Nav Relay": [["2%", "3%", "4%", "5%"], []],
-	"Insulated Engine Assembly": [
-		["100%", "10%", "50%"],
-		["100%", "90%"],
-	],
-	"Converted Hangar": [
-		[1.5, 1.5, 40, 20, 1, 5, 1],
-		["10%", "25%"],
-	],
-	"Defensive Targeting Array": [["50%"], [100]],
-	"Expanded Deck Crew": [["15%", "25%", 20], []],
-	"Recovery Shuttles": [["75%"], ["95%"]],
-	"Converted Fighter Bay": [[50, "20%", "80%"], ["15%"]],
-	"B-Deck": [["40%", "100%"], []],
-	"Fighter Chassis Storage": [[], []],
-	"Flux Coil Adjunct": [
-		[600, 1200, 1800, 3000],
-		[200, 400, 600, 1000],
-	],
-	"Flux Distributor": [
-		[30, 60, 90, 150],
-		[10, 20, 30, 50],
-	],
-	"Hardened Subsystems": [["50%", "25%"], []],
-	"Neural Interface": [[50], []],
-	"Neural Integrator": [[50, "10%"], []],
-	"Operations Center": [["250%"], []],
-	"Militarized Subsystems": [[1, "100%"], []],
-	"Phase Anchor": [[0, "2x", "1x"], []],
-	"Phase Field": [["50%", 5, 5], []],
-
-	"Expanded Cargo Holds": [[30, 60, 100, 200, "30%", "50%"], []],
-	"High Resolution Sensors": [
-		[50, 75, 100, 150, "10%"],
-		[1000, 1500, 2000, 2500],
-	],
-	"Efficiency Overhaul": [["20%", "50%"], ["10%"]],
-	"Surveying Equipment": [[5, 10, 20, 40, 5], ["100%"]],
-	"Shielded Cargo Holds": [[], []],
-	"Salvage Gantry": [["10%", "25%", "30%", "40%", "20%"], []],
-	"Civilian-grade Hull": [[100, 50], []],
-	"High Maintenance": [[100], []],
-};
-export default HULLMODS_DATA;
